Update asciid.velocity
Changed some basic stuff in the Macro description. I changed the name to `adoc` based on a suggestion by the Asciidoctor founder @mojavelinux. 

It makes sense because most folks who know AsciiDoc will be used to the .adoc prefix on the AsciiDoc files. I don't see a need to change any of the other parts of the macro code to strip out the `asciid` name you chose. Just the user-facing way of interacting with the macro would be a good change to make.